+++ packages/js/openreceive/src/angular-checkout-component.d.ts
@@ -1,1 +1,0 @@
-export * from "@openreceive/angular/checkout-component";