+++ packages/js/openreceive/src/vue-checkout.d.ts
@@ -1,1 +1,0 @@
-export { default } from "@openreceive/vue/checkout.vue";